requirements.txt
fastapi==0.135.1
sqlalchemy==2.0.48
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uvicorn[standard]==0.24.0
bcrypt==4.0.1
